requirements.txt
appdirs==1.4.4
asgiref==3.4.1
attrs==19.3.0
black==19.10b0
cached-property==1.5.2
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==2.0.3
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
defusedxml==0.7.1
Django==3.0.4
django-allauth==0.41.0
django-environ==0.4.5
django-rest-auth==0.9.5
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
drf-spectacular==0.17.2
drf-yasg==1.17.1
efficientnet==1.0.0
ExifRead==2.3.2
flake8==3.8.3
h5py==3.3.0
idna==3.2
image-classifiers==1.0.0
imageio==2.9.0
importlib-metadata==1.7.0
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
jsonfield==3.1.0
jsonschema==3.2.0
Keras-Applications==1.0.8
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.4.3
mccabe==0.6.1
networkx==2.6.2
numpy==1.21.2
oauthlib==3.1.1
opencv-python==4.5.3.56
packaging==20.4
pathspec==0.8.0
Pillow==8.3.1
progressbar2==3.53.1
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-utils==2.5.6
python3-openid==3.2.0
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
regex==2020.7.14
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scikit-image==0.18.2
scipy==1.7.1
segmentation-models==1.0.1
six==1.16.0
smartcrop==0.3.3
sqlparse==0.4.1
tifffile==2021.7.2
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.10.0.0
Unipath==1.1
uritemplate==3.0.1
urllib3==1.26.6
zipp==3.1.0
